Fixed the error where the player could play the game again after quiting from the pause menu or after losing.
Still have a bug where if the player presses a different key to 'space' in the pre-game screen that the game will freeze.

Have some optimazation to do still so the game runs smoothly.

Have to wait for me to make the file stream code for the CORE library so I can load different brick structures, save a leaderboard and titles from files.